refactor(installer): clarify DDOT path helpers and version layout
Split agent root vs ext/ddot path lookup with (string, error), move
agentVersionForExtensions and getCurrentAgentVersion out of the shared
extensions file, and drop the unused ddotExtensionPath wrapper.
